+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 ## Unreleased
 - Update to version 9.4
-- Added support to multiple return types on casting.
+- Added support to multiple return types on casting (fixes #196)
 - Implemented Subtype protocol to 10 generic types missing.
 - Added credo check to make sure new generic types don't miss the protocol implementation
+- Added support to media input files (fixes #195)
 
 ## [0.57.0]
 - Update to Elixir 1.19 and Erlang 28.1.1